requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
importlib-metadata==6.6.0
Markdown==3.4.3
Pillow==9.5.0
psycopg2==2.9.6
pytz==2023.3
redis==4.5.5
sqlparse==0.4.4
zipp==3.15.0
